+++ client/src/lib/locales/en-US.js
@@ -35,7 +35,6 @@ export default {
   nav: {
     home: 'Home',
     problems: 'Problems',
-    contribute: 'Contribute',
     leaderboard: 'Leaderboard',
     chain: 'Chain',
     login: 'Log in',